Fix array-put-object
Change CanPutArrayElementFromCode to take an object and a class, rather
than two classes.  This allows the generated code to skip the null check
on the element (I believe this was the plan all along, but I must have
forgotten and the current codegen was dereferencing the element to find it's
class without checking for NULL).

Change-Id: I11865e0e625ab0e800718398fd911733718e6ce9