Support for currently untested exceptions/errors.
Replace place holder implementations of exception/error code with code
that is likely mostly correct. Testing and creating detail messages are
TODO items for each of these exceptions/errors and warning are created
to reflect this need.

As we don't currently generate code to throw these exceptions, they are
untested.

Change-Id: I75876a9cca37892fa065eccdf34f9868b4142ec9