Reg utility fix
Long ago in a galaxy far away, there was a trace compiler that
handled short code fragments with a small, sparse and variable
set of temp registers.  In that situation, doing linear reg
lookups wasn't terrible.  In the new world that has a fixed and
packed set of registers, a linear search is unnecessary, foolish
and wasteful.  [P.S. perf showed that roughly 25% of all
compilation time was spent doing register manipulation!]

Change-Id: I6f23d9f70367fb4139cc28a27fd9fdf8beffa270