Finish the OutOfMemoryError implementation.
This copes with the double-OOME case.

Also check that I don't leave local references in a newly-attached thread's
local reference table, and fix the leaks this discovered.

Also fix the code that implements fillInNativeStackTrace to cope with the
situation where we're not able to allocate (because we're throwing
OutOfMemoryError and there's not enough space left for our arrays).

Also fix the order of checking for a pending exception and popping the
local references in the JNI native method invocation stub. (This fixes
the warnings we'd been seeing from the IndirectReferenceTable in test 064.)

Also improve some -Xcheck:jni output.

This fixes test 061.

Change-Id: Icc04a2c06339bd28d6772190350a86abfc5734b8