Hprof code cleanup.
This replaces a bunch of structs+functions with C++ objects
and member functions. It shortens names where unambiguous and
reduces visibility where plausible.

The main two changes are to create Hprof and HprofRecord classes
(from the hprof_record_t and hprof_context_t structs) and moves
functions into the classes. It uses stack allocation for the Hprof
object, which saves a few allocation-failure checks. And it uses
constructors and destructors rather than functions that do those
things.

This change should not alter behavior.

Change-Id: I62ce97fae3c1ba6e2ea89fdfdca915b013072868