Fix a couple of JDWP bugs.
First, dalvikvm's behavior of just carrying on if it failed to set up JDWP was
the opposite of helpful. It's wasted loads of my time lately, and I've already
benefitted in the half hour since making this a hard stop.

Second, I was accidentally reporting upcall frames. I don't know why I wasn't
always seeing this -- because we always make an upcall to 'main' -- but it was
repeatable if the stack was short enough.

Change-Id: Iee14867f596436bed5ca0546db71e62b62d11f41