Fix basic block construction
There was a bug in the CFG building mechanism that tiggered when
a backwards branch re-entered the current basic block at a point
other than the head.  In that case, we'd split the original basic block
in two.  However, the head of the old block then became the head
of the first of a new two block pair - and we really needed to continue
processing with the second of the pair.

Change-Id: Ieb9fdd5176d86cd0959bb9e91db9ca06075d4f21