Fill out the list of JDWP commands.
This also changes to reporting ERR_NOT_IMPLEMENTED rather than aborting for
a command we don't have a handler for. (As opposed to a command we have a
handler for, but whose handler just calls UNIMPLEMENTED(FATAL) --- there's
no change to that. This just means that for stuff like jdb's "threadlocks"
command that dalvik never supported, using it now causes jdb to say
"Command 'threadlocks' is not supported on the target VM" instead of causing
the VM to abort and jdb to be thereby disconnected.)

Change-Id: I1b7952d4ca6a540afb9fde6ae54de11b36345d4e