Improve the ThreadStress test to send SIGQUIT, and fix the bug that finds.
I'd heard complaints of a system_server deadlock on crespo, but wasn't seeing
it on mysid. I did see it soon after trying to use crespo, but system_server
has way too many threads and adb is too damn unreliable for me to effectively
debug that. This improves ThreadStress so we exercise that path lots, and
catch the deadlock relatively quickly with relatively few threads.

The fix is explained in a code comment that's larger than the fix.

Change-Id: I593bc94cf1239065a604703568420986a03ce628