Mac runtime fixes.
This is enough to let us run test-art-host on the Mac. The tests
that don't start a runtime (and thus don't try to dlopen(3)
libjavacore.dylib) succeed. The others fail, but I'm not sure how
the Linux tests find libjavacore.so, so I'm not sure how to fix
things for Mac OS.

Change-Id: Idf489fc09f0a8f64fbb60dfcdbca9c3df2d4fdcb