Fix a few bugs in push-and-run-test-jar.
Simon says it's "adb shell mkdir", not "adb mkdir" (which was silently failing).

We shouldn't be putting test .jar and .oat files in /system/framework (silently
causing --quiet to use out of date code).

The runtime has been looking for X.jar.oat given X.jar for some time now (which
was silently causing all modes to use out of date code in conjunction with a
to-be-fixed bug that's preventing us from recognizing that the .oat file's
checksum is out of date).

Change-Id: I2421878159f24cc7564faa105a3e795c09929158