x86 codegen source file skeletons
Add source files for x86 code generation.  These are basically just
the MIPS versions with all bodies ifdef'd out and replaced with
UNIMPLEMENTED(WARNING).

There is also one real change in this CL related to the way the
assembler computes instruction size.  We now set the initial
instruction size into the LIR during flag generation.  For x86, this
size may be changed during assembly when we detect the ability to
substitute a general instruction for an equivalent one with a shorter
encoding.

NOTE: the new files are not yet ready for compilation.  That will be
the next CL.

Change-Id: Id94fbeeb7c3e36ba4a8d93c3807351ea20c74133