Fix "wide" cts vm-tests
In change https://android-git.corp.google.com/g/#/c/171904/ I
removed a filler word that was placed in the gap between
the last of the Dalvik locals and the first of the ins.  This
filler existed to allow use of the last local as a double - something
that dx will never do (and I managed to convince myself that it
would never happen).  However, it did.  Some of the cts tests use
directly generated Dalvik code, and a group of them used the
bad pattern.  We could update the verifier to check for and
explicitly disallow this.  But, in case there is some other non-test
code out there which uses this pattern it's probably best to just
put the filler word back.

Yuck.

Change-Id: I67e6049f09007d3996ab6165c8160c56e4948600