Emit the ELF image into the memory buffer.
Emit the ELF image to the memory buffer, so that we can copy
and embed them into Oat file in the future.  Currently, We
are keeping writing the ELF image to the file, such as
gtest-0.elf, for the unit test.  However, we will remove them
after the Oat file related commits are ready.

Change-Id: Ic961bd31a160df69d296b05e3f83bf85b17b031f