Do not use FindOatFileFromOatLocation in DexFile_isDexOptNeeded
isDexOptNeeded is trying to sniff out bad files and
FindOatFileFromOatLocation will register the file on the opened oat
file list so later if the same process tries to run code from one of
the oats, it will find it already opened and not validate its contents
are up-to-date. This happens in the special case of the system server
which is responsible for scanning to make sure things are up-to-date
proactively, but also starts up services using those files.

Change-Id: Ia9b483a46336b46bc1ec22180e60099e74886927