Remove unnecessary zero-initializer.
Dex verifier will guarantee that a Dalvik register will be
initialized before any usage.  So we don't have to generate
the code to initialize them to 0 after allocation.

NOTE: We wrote the zero-initializer for historical reason.
There was a bug in the implementation of SHL, SHR, USHR,
and result in undef value after optimization.  Since we
have found the bug, and we don't get undef any more, it's
time to remove these zero-initializer to save code size
and run time.

Change-Id: I0375af3bed08495f6743699c7a72c295cad71dfa