Don't pre-allocate one OOME per thread.
There's no associated stack trace, so we can share one between all
threads, saving a little memory and thread start-up time. Speaking
of stack traces: dump the stack to the log at the point where we
realize we're not going to be able to throw an exception with a
stack, so the developer has _something_ to work with.

Change-Id: I2246d291855bd9b9ee13f2be5b1ce14f669e9410