Quick compiler: fix for missing defs
When the verifier replaces a Dex instruction, it may cause us
to lose a definition, making the SSA graph invalid.  This CL
deals with this situation by introducing dummy definitions.

With this CL, run-tests 042 and 075 pass.  The only remaining
run-test failures are related to missing exception edges in the CFG.
The old compiler could infer some of them, while llvm requires all
to be explicit.

Change-Id: I471fa515d8717aa102713a05b0e4115bf656b78e