Cleanup gbc_expander for fly2iceland.
Add the to-be-implemented cases to the expander.

Fix bug: Should get the RHS with call_inst.getArgOperand(1).

Update comments in intrinsics definition to fit the code.

Reorder the switch cases so that every instrinsics will
be in the same order with the instrinsics_func_list.def.

Implement integer shift expansion for GBC.

Change-Id: I8b3a848d5627445a2cef6e1675579b7fd0ca89bb