Fix image size regression
The new basic block collapse pass marks the eliminated block as
kDead.  On the Quick side, we skip code generation for these blocks - but
neglected to do so for the non-Quick build.  The result: lots of
extra useless unreachable code got generated - hence the image size
bloat.

Fixed, and good news is that we're now seeing the roughly the same
decrease in image size that showed up in the Quick build.

Change-Id: Ie8038cb8058020eefaae1085362a2f8345895284