Compiler: Take advantage of constant propagation
The common frontend tracks constants via a constant propagation pass.
When converting from MIR to GBC (for Portable) or LIR (for Quick),
recognize constant arguments and select more efficient codegen forms.

Note: we still have to flush constants to their associated vregs to
support deoptimization.  There's quite a bit of possible code size
gain if we were to eliminate unnecessary stores or enhance the vmap
table to explicitly represent the ranges over which Dalvik vregs
are constant.

Also some minor code refactoring related to array operations.  There
are sufficient architectural differences to make it worthwhile to
have target-dependent aget/aput generators.  On Arm, this is mostly
beneficial to floating point array loads and stores.

This CL yields a ~0.8% decrease in code size over the framework,
and a nice pop to a few of the standard point benchmarks
(linpack: ~10%, cm: ~11%, scimark: ~13% - no significant change to
the others)

Change-Id: I2337e1aa0622b34a34c3775f8b7dbf5e6969da3e