Implement EnableCollection, DisableCollection, IsCollected, and fix Exit.
Rewrites the object registry to not just be one big leak. We use jweaks
by default, which means (a) the GC doesn't need to know anything about us
and (b) we don't cause unnecessary heap retention. We promote to regular
JNI global references if the debugger asks us to.

The problem with VirtualMachine.Exit turned out to be that you're supposed
to send a success reply to the command before exiting. This is a bit awkward
with our current division of responsibilities, but it lets us pass another
test.

Also log a summary of our replies when -verbose:jdwp is in effect, not
just the requests.

Change-Id: Idb33e99fe7d8bee7a79152d81fee42e2af00852b